Move more windows stuff into windows folder
