+++ trview.app/Menus/ViewMenu.cpp
@@ -1,5 +1,5 @@
 #include "ViewMenu.h"
-#include <trview.app/WindowIDs.h>
+#include <trview.app/Windows/WindowIDs.h>
 
 namespace trview
 {
