+++ trview.app/Windows/WindowResizer.cpp
