+++ trview.app/Windows/WindowResizer.h
