+++ trview.app/trview.app.vcxproj.filters
@@ -1,7 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <ClCompile Include="WindowResizer.cpp" />
     <ClCompile Include="Item.cpp">
       <Filter>Items</Filter>
     </ClCompile>
@@ -141,10 +140,11 @@
     <ClCompile Include="Menus\ViewMenu.cpp">
       <Filter>Menus</Filter>
     </ClCompile>
+    <ClCompile Include="Windows\WindowResizer.cpp">
+      <Filter>Windows</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="WindowResizer.h" />
-    <ClInclude Include="WindowIDs.h" />
     <ClInclude Include="Item.h">
       <Filter>Items</Filter>
     </ClInclude>
@@ -296,6 +296,12 @@
     </ClInclude>
     <ClInclude Include="Menus\ViewMenu.h">
       <Filter>Menus</Filter>
+    </ClInclude>
+    <ClInclude Include="Windows\WindowIDs.h">
+      <Filter>Windows</Filter>
+    </ClInclude>
+    <ClInclude Include="Windows\WindowResizer.h">
+      <Filter>Windows</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
